+++ src/Api/CurrentUser/Workspaces/Permissions/AbstractPermissionsApi.php
@@ -1,0 +1,25 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Bitbucket API Client.
+ *
+ * (c) Graham Campbell <hello@gjcampbell.co.uk>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Bitbucket\Api\CurrentUser\Workspaces\Permissions;
+
+use Bitbucket\Api\CurrentUser\Workspaces\AbstractWorkspacesApi;
+
+/**
+ * The abstract permissions API class.
+ *
+ * @author Graham Campbell <hello@gjcampbell.co.uk>
+ */
+abstract class AbstractPermissionsApi extends AbstractWorkspacesApi
+{
+}