+++ CHANGELOG.md
@@ -6,7 +6,8 @@ CHANGE LOG
 
 * Add PHP 8.5 support
 * Fixed the current user workspaces endpoint
-* Added the current user workspace permission endpoint
+* Added the current user workspaces API
+* Deprecated current user workspaces pass-through methods
 
 
 ## V5.0 (23/02/2025)